Add files via upload
In this project, various machine learning algorithms and strategies are applied to analyze the search trends
of symptoms and to predict the hospitalization cases based on their patterns. Two datasets respectively on
the search trends of symptoms and the hospitalization cases were cleaned and merged for use. Visualization,
clustering, and dimension reduction were performed on the search trends of symptoms to understand the feature
data clearly. k-nearest neighbors algorithm and decision tree algorithm are applied to train the models for
predicting the hospitalization cases based on the processed dataset. Their regression performances are compared
and analyzed. The result shows that using decision tree on the whole dataset achieves the best regression
performance for this dataset and is generally fast to train. The result also shows that the data from one region is
less solid for making predictions for other regions. Finally, we concluded that the k-nearest neighbors algorithm
is suitable for datasets with closely correlated data, whereas the decision tree algorithm excels for datasets of
larger size.